requirement.txt
absl-py==1.0.0
anyio==3.5.0
argcomplete==1.10.3
asgiref==3.5.0
astunparse==1.6.3
autopep8==1.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.8.2
cachetools==5.0.0
category-encoders==2.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
compressed-rtf==1.0.6
cryptography==36.0.1
cycler==0.11.0
decorator==5.1.1
Django==4.0.2
django-environ==0.8.1
django-filter==21.1
djangorestframework==3.13.1
docx2txt==0.8
ebcdic==1.1.1
Elixir==0.7.1
et-xmlfile==1.1.0
extract-msg==0.28.7
fastapi==0.74.1
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.44.0
h11==0.13.0
h5py==3.6.0
idna==3.3
IMAPClient==2.1.0
importlib-metadata==4.11.4
Jinja2==3.0.3
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
libclang==13.0.0
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib==3.5.1
nltk==3.7
numpy==1.22.2
oauthlib==3.2.0
olefile==0.46
openpyxl==3.0.9
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
patsy==0.5.2
pbr==5.8.1
pdfminer.six==20191110
Pillow==9.0.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
pycryptodome==3.14.1
pydantic==1.9.0
pyparsing==3.0.7
PyPDF2==1.26.0
python-dateutil==2.8.2
python-docx==0.8.11
python-pptx==0.6.21
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
regex==2022.1.18
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.12.0
sklearn==0.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
SQLAlchemy==0.7.10
sqlalchemy-migrate==0.11.0
sqlparse==0.4.2
starlette==0.17.1
statsmodels==0.13.2
Tempita==0.5.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
toml==0.10.2
tqdm==4.62.3
typing-extensions==4.1.1
tzdata==2021.5
tzlocal==4.2
urllib3==1.26.8
uvicorn==0.17.5
Werkzeug==2.0.3
wrapt==1.13.3
xlrd==1.2.0
XlsxWriter==3.0.3
xlwt==0.7.2
zipp==3.8.0
